add apply_filter_from and remove_filter_from methods

+++ crates/png-glitch/examples/transpose.rs
@@ -4,7 +4,7 @@ fn main() {
     let mut glitch = png_glitch::PngGlitch::open("etc/sample00.png").unwrap();
     glitch.remove_filter();
 
-    let src = glitch.height() as usize / 3;
+    let src = glitch.height()/ 3;
     let dest = src * 2;
     let width = glitch.height() / 10;
